// database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  public: {
    Tables: {
// ...
      logs: {
        Row: {
          archetype: string | null
          created_at: string
          decklist_id: string | null
          id: string
          log: string
          notes: string | null
          opp_archetype: string | null
          result: string | null
          turn_order: string | null
          user: string
          format: string
        }
        Insert: {
          archetype?: string | null
          created_at?: string
          decklist_id?: string | null
          id?: string
          log: string
          notes?: string | null
          opp_archetype?: string | null
          result?: string | null
          turn_order?: string | null
          user: string
          format: string
        }
        Update: {
          archetype?: string | null
          created_at?: string
          decklist_id?: string | null
          id?: string
          log?: string
          notes?: string | null
          opp_archetype?: string | null
          result?: string | null
          turn_order?: string | null
          user?: string
          format: string
        }
        Relationships: [
          {
            foreignKeyName: "logs_user_fkey"
            columns: ["user"]
            isOneToOne: false
            referencedRelation: "top_5_users_battle_logs"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      decklists: {
        Row: {
          archetype: string | null
          cards: Json
          content_hash: string | null
          created_at: string
          format: string | null
          game: string
          id: string
          name: string
          updated_at: string
          user_id: string
        }
        Insert: {
          archetype?: string | null
          cards: Json
          content_hash?: string | null
          created_at?: string
          format?: string | null
          game?: string
          id?: string
          name: string
          updated_at?: string
          user_id: string
        }
        Update: {
          archetype?: string | null
          cards?: Json
          content_hash?: string | null
          created_at?: string
          format?: string | null
          game?: string
          id?: string
          name?: string
          updated_at?: string
          user_id?: string
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}